Avoid copies on calls to parser.Sources()
This is passing around a `map[string][]byte`, which is fine, but the
underlying implementation makes a copy every time it's called, and it
doesn't appropriately pre-allocate the size of the map to make that more
efficient. This generates a lot of garbage, especially given that
nothing modifies what is returned by Sources().

Instead, we can call the `Files()` method, which returns a
`map[string]*hcl.File`, which still gives us access to the underlying
`[]byte` via the `Bytes` field, but doesn't require making a copy of the
map (or modifying upstream to fix this).

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>